Week 6: Solution - Hibernate&SpringORM

+++ src/main/java/com/github/blog/dao/OrderDao.java
@@ -5,5 +5,5 @@ import com.github.blog.model.Order;
 /**
  * @author Raman Haurylau
  */
-public interface OrderDao extends CrudDao<Order, Integer> {
+public interface OrderDao extends CrudDao<Order, Long> {
 }
