+++ src/main/java/com/github/blog/dao/PostReactionDao.java
@@ -5,5 +5,5 @@ import com.github.blog.model.PostReaction;
 /**
  * @author Raman Haurylau
  */
-public interface PostReactionDao extends CrudDao<PostReaction, Integer> {
+public interface PostReactionDao extends CrudDao<PostReaction, Long> {
 }
