+++ src/main/java/com/github/blog/dao/TagDao.java
@@ -5,5 +5,5 @@ import com.github.blog.model.Tag;
 /**
  * @author Raman Haurylau
  */
-public interface TagDao extends CrudDao<Tag, Integer> {
+public interface TagDao extends CrudDao<Tag, Long> {
 }
